introduce output module in machine
