+++ kv_store/tests/fixtures/mod.rs
@@ -1,4 +1,5 @@
-use edfsm::{Change, Drain, Fsm, Init, Terminating};
+use edfsm::{Change, Fsm, Terminating};
+use machine::output::OutputBuffer;
 use serde::{Deserialize, Serialize};
 
 pub struct Counter;
@@ -61,36 +62,6 @@ impl Fsm for Counter {
             se.push(Output::Tock);
         }
     }
-}
-
-#[derive(Debug)]
-pub struct OutputBuffer<A>(pub std::vec::Vec<A>);
-
-impl<A> OutputBuffer<A> {
-    pub fn push(&mut self, item: A) {
-        self.0.push(item);
-    }
-}
-
-impl<A> Default for OutputBuffer<A> {
-    fn default() -> Self {
-        Self(Vec::new())
-    }
-}
-
-impl<A> Drain for OutputBuffer<A>
-where
-    A: Send,
-{
-    type Item = A;
-
-    fn drain_all(&mut self) -> impl Iterator<Item = Self::Item> {
-        self.0.drain(0..)
-    }
-}
-
-impl<S, A> Init<S> for OutputBuffer<A> {
-    fn init(&mut self, _: &S) {}
 }
 
 impl Terminating for Event {
